move folder; change example to a more general agent
